requirements.txt
# AI Kit 依赖包

# 核心依赖
python-dotenv>=1.0.0
click>=8.0.0
pydantic>=2.5.0

# HTTP 相关
h11>=0.14.0
httpcore>=0.17.0
httpx>=0.25.0

# OpenAI/LangChain 相关
openai>=1.30.0
langchain-openai>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-text-splitters>=0.2.0

# 向量数据库
chromadb>=0.4.0

# 网页解析
beautifulsoup4>=4.12.0

# 开发依赖
pytest>=7.0.0
pytest-cov>=4.0.0